Mobile_Python/requirements.txt
behave==1.2.6
Appium-Python-Client==2.1.0
selenium==4.1.0
allure-behave==2.8.16